pop22_cohort

A transformed and summarized population data, distinguishing between Estimate and Projections, reformatted for clarity, and adding a state-wide summary.

Usage

pop22_cohort

Format

Data frame including FIPS, County, Year, Type, Gender and Population.

AGE GROUP

Age Group of Population, "0 TO 4" "5 TO 9" "10 TO 14" "15 To 19" "20 TO 24" "25 TO 29" "30 TO 34" "35 TO 39" "40 TO 44" "45 TO 49" "50 TO 54" "55 TO 59" "60 TO 54" "65 TO 69" "70 TO 74" "75 TO 79" "80 TO 84" "85+"

FIPS

FIPS code of each county

County

Name of county

Year

Year of Projection/Estimate (2010-2038)

Type

Estimate or Projection

Gender

Male , Female or Total

Population

Population number

Examples

data(pop22_cohort)
head(pop22_cohort)
#> # A tibble: 6 × 7
#>   `AGE GROUP` County     FIPS Gender  Year Population Type    
#>   <chr>       <chr>     <dbl> <chr>  <int>      <dbl> <chr>   
#> 1 0 TO 4      ABBEVILLE     1 Total   2010       1503 Estimate
#> 2 0 TO 4      ABBEVILLE     1 Female  2010        719 Estimate
#> 3 0 TO 4      ABBEVILLE     1 Male    2010        784 Estimate
#> 4 0 TO 4      ABBEVILLE     1 Total   2011       1436 Estimate
#> 5 0 TO 4      ABBEVILLE     1 Female  2011        702 Estimate
#> 6 0 TO 4      ABBEVILLE     1 Male    2011        734 Estimate