pop_proj Results of extending the SC RFA projections This table includes detailed results of the methods used to extend the projections (for both editions of the SC RFA projections).

Format

A data frame with columns:

Period

Original (from SC RFA) or Extended (from the methods used to extend the projections)

Scenario

Estimate, Moderate, or High

Edition

2019 or 2022

County

Name of the county

Year

Year of Projection (2000-2070)

Type

Estimate, SC RFA Projection, or Extrapolated Projection

Population

Population number

T_0

Initial year for this "scenario"

T_1

Final year for this "scenario"

Pop_0

Initial population for this "scenario"

Pop_1

Final population for this "scenario"

Growth_linear

Population growth for this "scenario," annual change in population

Growth_exp

Exponential growth rate for this "scenario"

Growth_moderate

Annual change in population under the moderate growth scenario (linear growth with minimum set to zero)

Growth_exp_floor

Minimum growth rate for the High scenario (the state-wide average*1.1)

Growth_hi

Exponential growth rate for this "scenario," floor set to at least the state average growth rate

Type2

Relative growth in population under the moderate growth scenario (2019 edition)

pop_proj pop_proj Results of extending the SC RFA projections This table includes detailed results of the methods used to extend the projections (for both editions of the SC RFA projections). data(pop_proj) head(pop_proj) datasets