pop_proj_relative: Relative Growth Projections for Different Counties

This dataset provides relative growth projections for different counties under moderate and high growth scenarios. It includes calculations for growth rates and driver growth factors under each scenario.

Usage

pop_proj_relative

Format

The dataset is a data frame with the following columns:

County

Name of the county

Year

Year of Projection (2000-2038)

High

Population in the high growth scenario

Moderate

Population in the moderate growth scenario

Driver_growth_mod

Relative growth in population under the moderate growth scenario

Driver_growth_hi

Relative growth in population under the high growth scenario

The formulas used in the dataset are as follows:

Moderate Growth Scenario (Moderate): $$Growth\_{moderate} = \max(Growth\_{linear}, 0)$$ Here, \(Growth\_{linear}\) represents the average annual increase in population.

High Growth Scenario (High): $$Growth\_{hi} = Growth\_{exp\_floor} \times 1.10$$ \(Growth\_{exp}\) is the exponential growth rate. \(Growth\_{exp\_floor}\) is set to at least the state average growth rate.

Driver Growth Factors: $$Driver\_{growth\_mod} = \frac{Moderate Population\ in\ Year\ X}{Moderate Population\ in\ 2021}$$ $$Driver\_{growth\_hi} = \frac{High Population\ in\ Year\ X}{High Population\ in\ 2021}$$

Examples

data(pop_proj_relative)
head(pop_proj_relative)
#> # A tibble: 6 × 6
#>   County     Year   High Moderate Driver_growth_mod Driver_growth_hi
#>   <chr>     <int>  <dbl>    <dbl>             <dbl>            <dbl>
#> 1 ABBEVILLE  2021 24071.   24071.             1                 1   
#> 2 ABBEVILLE  2022 24291.   23930.             0.994             1.01
#> 3 ABBEVILLE  2023 24513.   23789.             0.988             1.02
#> 4 ABBEVILLE  2024 24738.   23648.             0.982             1.03
#> 5 ABBEVILLE  2025 24964.   23507.             0.977             1.04
#> 6 ABBEVILLE  2026 25192.   23345.             0.970             1.05